+++ test.py
@@ -1,33 +1,34 @@
-from heap import *
+from minheap import MinHeap
+from graph import Graph, Vertex
 
 
 def testGraph():
     g = Graph()
-    for i in range(100):
-        g.addVertex(Vertex(name=i, dist=100-i))
+    for i in range(10):
+        g.addVertex(Vertex(name=i, distance=100-i))
     print(g)
     g.show()
 
 
 def testHeap(heap=None):
     h = heap or MinHeap()
-    V = [Vertex(name=i, dist=10-i) for i in range(10)]
+    V = [Vertex(name=i, distance=10-i) for i in range(10)]
     for v in V:
-        h.push(v)
+        h.insert(v)
     h.show()
 
     for v in V:
-        h.decreaseKey(v, v.dist-10)
+        h.decreaseKey(v, v.distance-10)
         h.show()
 
 
 def testDijkstra(graph=None, heap=None, show=False):
     h = heap or MinHeap()
     g = graph or Graph()
-    for i in range(100):
-        g.addVertex(Vertex(name=i, dist=i))
-    for i in range(100):
-        g.addEdge(u=i, v=(i+1) % 100, w=2)
+    for i in range(10):
+        g.addVertex(Vertex(name=i, distance=i))
+    for i in range(10):
+        g.addEdge(u=i, v=(i+1) % 10, w=2) # make linear cycle graph
 
     if show:
         g.show()
@@ -38,7 +39,7 @@ def testDijkstra(graph=None, heap=None, show=False):
 if __name__ == '__main__':
 
     # testHeap(HeapqHeap())
-    testHeap()
-    # testDijkstra(show=True)
+    # testHeap()
+    testDijkstra(show=True)
 
     # testGraph()
